+++ tests/string/trimSuffix.js
@@ -1,0 +1,18 @@
+import trimSuffix from './../../src/string/trimSuffix.js'
+
+describe('trimSuffix', () => {
+  it('', () => {
+    const string = '/es/test/es/page/es/'
+
+    expect(trimSuffix(string, '/')).toEqual('/es/test/es/page/es')
+    expect(trimSuffix(string, '/es/')).toEqual('/es/test/es/page')
+    expect(trimSuffix(string, '/es/page/es/')).toEqual('/es/test')
+    expect(trimSuffix(string, '/es/test/es/page/es/')).toEqual('')
+
+    expect(trimSuffix(string, '/es/test/es/page/es//')).toEqual(string)
+    expect(trimSuffix(string, '//es/test/es/page/es/')).toEqual(string)
+    expect(trimSuffix(string, '')).toEqual(string)
+    expect(trimSuffix(string, '/es')).toEqual(string)
+    expect(trimSuffix(string)).toEqual(string)
+  })
+})